Refactor(): Observer pattern + refactor
Moved TeamService to observer pattern, added observer pattern for team
