+++ src/main/java/ecostruxure/rate/calculator/bll/profile/IProfileObserver.java
@@ -2,6 +2,6 @@ package ecostruxure.rate.calculator.bll.profile;
 
 import ecostruxure.rate.calculator.be.Profile;
 
-public interface ProfileObserver {
+public interface IProfileObserver {
     void update(Profile profile);
 }
