+++ src/main/java/ecostruxure/rate/calculator/bll/profile/observers/TeamTotalObserverI.java
@@ -3,7 +3,7 @@ package ecostruxure.rate.calculator.bll.profile.observers;
 import ecostruxure.rate.calculator.be.Profile;
 import ecostruxure.rate.calculator.be.Team;
 import ecostruxure.rate.calculator.be.TeamProfile;
-import ecostruxure.rate.calculator.bll.profile.ProfileObserver;
+import ecostruxure.rate.calculator.bll.profile.IProfileObserver;
 import ecostruxure.rate.calculator.dal.ITeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 @Service
-public class TeamTotalObserver implements ProfileObserver {
+public class TeamTotalObserverI implements IProfileObserver {
     @Autowired
     private ITeamRepository teamRepository;
