+++ src/main/java/ecostruxure/rate/calculator/bll/team/ITeamObserver.java
@@ -1,0 +1,7 @@
+package ecostruxure.rate.calculator.bll.team;
+
+import ecostruxure.rate.calculator.be.Team;
+
+public interface ITeamObserver {
+    void update(Team team);
+}
