+++ src/main/java/ecostruxure/rate/calculator/bll/team/observers/ProjectObserver.java
@@ -1,0 +1,4 @@
+package ecostruxure.rate.calculator.bll.team.observers;
+
+public class ProjectObserver {
+}
