+++ src/main/java/ecostruxure/rate/calculator/controllers/TeamProfileController.java
@@ -1,14 +1,6 @@
 package ecostruxure.rate.calculator.controllers;
 
-import ecostruxure.rate.calculator.be.Team;
-import ecostruxure.rate.calculator.be.TeamProfile;
-import ecostruxure.rate.calculator.be.dto.TeamDTO;
-import ecostruxure.rate.calculator.bll.service.TeamService;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/team-profile")
