+++ src/main/java/ecostruxure/rate/calculator/controllers/TeamsController.java
@@ -3,7 +3,7 @@ package ecostruxure.rate.calculator.controllers;
 import ecostruxure.rate.calculator.be.Team;
 import ecostruxure.rate.calculator.be.TeamProfile;
 import ecostruxure.rate.calculator.be.dto.TeamDTO;
-import ecostruxure.rate.calculator.bll.service.TeamService;
+import ecostruxure.rate.calculator.bll.team.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
